feat(db): 数据库从 SQLite 迁移到 Supabase Postgres
- 新增 schema.pg.sql (SERIAL + BIGINT 时间戳)
- 重写 db.js: 用 pg 直连, 兼容层保持原有 prepare/get/all/run 接口
  * ? 占位符自动转 $1..$n
  * INSERT 自动追加 RETURNING id 补 lastInsertRowid
  * int8 类型转 number, 避免时间戳变成字符串
  * 未配置 DATABASE_URL 时明确报错, 不再退化内存库
- SQL 方言: COUNT(*) -> COUNT(*)::int, MAX(0,x) -> GREATEST(0,x)
- app.js 加 catchAsync 包裹路由, 修复 async 异常导致 Node 进程退出
- render.yaml 环境变量 TURSO_* -> DATABASE_URL

+++ server/src/routes/posts.js
@@ -55,7 +55,7 @@ router.get('/', async (req, res) => {
   }
   const whereSql = where.length ? `WHERE ${where.join(' AND ')}` : '';
 
-  const total = (await db.prepare(`SELECT COUNT(*) AS c FROM posts p ${whereSql}`).get(...params)).c;
+  const total = (await db.prepare(`SELECT COUNT(*)::int AS c FROM posts p ${whereSql}`).get(...params)).c;
 
   const orderBy = sort === 'hot'
     ? 'ORDER BY (p.like_count + p.reply_count) DESC, p.created_at DESC'
